pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "road-topology"
version = "0.1.0"
description = "Road Topology Segmentation with YOLOv8 and SAM"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Road Topology Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "boxmot>=10.0.0",
    "segment-anything>=1.0",
    "transformers>=4.30.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=9.5.0",
    "albumentations>=1.3.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "structlog>=23.0.0",
    "pydantic>=2.0.0",
    "einops>=0.6.0",
    "timm>=0.9.0",
    "scipy>=1.10.0",
    "scikit-image>=0.21.0",
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0;sys_platform=='darwin'",
    "onnxruntime-gpu>=1.15.0;sys_platform!='darwin'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.3.0",
    "ruff>=0.0.270",
]

[project.scripts]
road-topo = "road_topology.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=road_topology --cov-report=term-missing --cov-report=html"
